+++ CHANGES.md
@@ -31,6 +31,9 @@
 - Fixed a fault that sometimes caused the "Next" button in the "Search" dialog to remain disabled after selecting a history item.
 - Search history will now check for any case when adding new items (so Match Case searches work).
 
+#### **Views:**
+- You can now minimize/restore a Month/Day in the All Events/Weekly views (this is remembered between opens/closes).
+
 #### **New Custom Trigger Events:**
 - Added a new custom trigger "onVisibleGroupsChanged", which is triggered when the visible groups are changed (passes the visible group names).
 - Added a new custom trigger "onVisibleEventTypesChanged", which is triggered when the visible event types are changed (passes the visible event type IDs).